light-agent-v2/requirements.txt
strands-agents[otel]
strands-agents-tools
bedrock-agentcore
boto3
aws-opentelemetry-distro
